add trxdb and trxstorage interface
