+++ internal/pkg/chainsdk/core/group.go
@@ -10,7 +10,7 @@ import (
 	"github.com/rumsystem/quorum/internal/pkg/conn"
 	"github.com/rumsystem/quorum/internal/pkg/logging"
 	"github.com/rumsystem/quorum/internal/pkg/nodectx"
-	"github.com/rumsystem/quorum/internal/pkg/storage"
+	"github.com/rumsystem/quorum/internal/pkg/storage/def"
 	quorumpb "github.com/rumsystem/rumchaindata/pkg/pb"
 	"google.golang.org/protobuf/proto"
 )
@@ -173,12 +173,12 @@ func (grp *Group) GetBlock(blockId string) (*quorumpb.Block, error) {
 
 func (grp *Group) GetTrx(trxId string) (*quorumpb.Trx, []int64, error) {
 	group_log.Debugf("<%s> GetTrx called", grp.Item.GroupId)
-	return nodectx.GetDbMgr().GetTrx(trxId, storage.Chain, grp.ChainCtx.nodename)
+	return nodectx.GetNodeCtx().GetChainStorage().GetTrx(trxId, def.Chain, grp.ChainCtx.nodename)
 }
 
 func (grp *Group) GetTrxFromCache(trxId string) (*quorumpb.Trx, []int64, error) {
 	group_log.Debugf("<%s> GetTrx called", grp.Item.GroupId)
-	return nodectx.GetDbMgr().GetTrx(trxId, storage.Cache, grp.ChainCtx.nodename)
+	return nodectx.GetNodeCtx().GetChainStorage().GetTrx(trxId, def.Cache, grp.ChainCtx.nodename)
 }
 
 func (grp *Group) GetProducers() ([]*quorumpb.ProducerItem, error) {
