+++ internal/pkg/storage/dbmgr.go
@@ -2,7 +2,6 @@ package storage
 
 import (
 	"errors"
-	"fmt"
 	"github.com/rumsystem/quorum/internal/pkg/logging"
 	"github.com/rumsystem/quorum/internal/pkg/utils"
 	quorumpb "github.com/rumsystem/rumchaindata/pkg/pb"
@@ -30,13 +29,6 @@ const ALLW_LIST_PREFIX string = "alw_list"     //allow list
 const DENY_LIST_PREFIX string = "dny_list"     //deny list
 const NONCE_PREFIX string = "nonce"            //group trx nonce
 const SNAPSHOT_PREFIX string = "snapshot"      //group snapshot
-
-type TrxStorageType uint
-
-const (
-	Chain TrxStorageType = iota
-	Cache
-)
 
 //groupinfo db
 const GROUPITEM_PREFIX string = "grpitem" //relay
@@ -109,69 +101,6 @@ func (dbMgr *DbMgr) TryMigration(nodeDataVer int) {
 			dbmgr_log.Errorf("db migration v1 for groupinfodb err %s", err)
 		}
 	}
-}
-
-//Get Trx
-func (dbMgr *DbMgr) GetTrx(trxId string, storagetype TrxStorageType, prefix ...string) (t *quorumpb.Trx, n []int64, err error) {
-	nodeprefix := utils.GetPrefix(prefix...)
-	var trx quorumpb.Trx
-	var nonces []int64
-
-	var key string
-	if storagetype == Chain {
-		key = nodeprefix + TRX_PREFIX + "_" + trxId
-		err = dbMgr.Db.PrefixForeach([]byte(key), func(k []byte, v []byte, err error) error {
-			if err != nil {
-				return err
-			}
-			perr := proto.Unmarshal(v, &trx)
-			if perr != nil {
-				return perr
-			}
-			nonces = append(nonces, trx.Nonce)
-			return nil
-		})
-		trx.StorageType = quorumpb.TrxStroageType_CHAIN
-	} else if storagetype == Cache {
-		key = nodeprefix + CHD_PREFIX + "_" + BLK_PREFIX
-		err = dbMgr.Db.PrefixForeach([]byte(key), func(k []byte, v []byte, err error) error {
-			if err != nil {
-				return err
-			}
-			chunk := quorumpb.BlockDbChunk{}
-			perr := proto.Unmarshal(v, &chunk)
-			if perr != nil {
-				return perr
-			}
-			if chunk.BlockItem != nil && chunk.BlockItem.Trxs != nil {
-				for _, blocktrx := range chunk.BlockItem.Trxs {
-					if blocktrx.TrxId == trxId {
-						nonces = append(nonces, blocktrx.Nonce)
-
-						clonedtrxbuff, _ := proto.Marshal(blocktrx)
-						perr = proto.Unmarshal(clonedtrxbuff, &trx)
-						if perr != nil {
-							return perr
-						}
-						trx.StorageType = quorumpb.TrxStroageType_CACHE
-						return nil
-					}
-				}
-			}
-
-			return nil
-		})
-
-	}
-
-	return &trx, nonces, err
-}
-
-func (dbMgr *DbMgr) IsTrxExist(trxId string, nonce int64, prefix ...string) (bool, error) {
-	nodeprefix := utils.GetPrefix(prefix...)
-	key := nodeprefix + TRX_PREFIX + "_" + trxId + "_" + fmt.Sprint(nonce)
-
-	return dbMgr.Db.IsExist([]byte(key))
 }
 
 //get block chunk
