+++ pkg/chainapi/appapi/handle.go
@@ -4,13 +4,13 @@ import (
 	"context"
 
 	"github.com/rumsystem/quorum/internal/pkg/appdata"
-	"github.com/rumsystem/quorum/internal/pkg/storage"
+	"github.com/rumsystem/quorum/internal/pkg/storage/def"
 )
 
 type Handler struct {
 	Ctx       context.Context
 	Appdb     *appdata.AppDb
-	Chaindb   *storage.DbMgr
+	Trxdb     def.TrxStorageIface
 	Apiroot   string
 	GitCommit string
 	ConfigDir string
